--- ninja-jobs/data/db.json
@@ -1,7 +1,0 @@
-{
-  "jobs": [
-    { "title": "Ninja UX Designer", "id": 1, "details": "lorem" },
-    { "title": "Ninja Web Developer", "id": 2, "details": "lorem" },
-    { "title": "Ninja Vue Developer", "id": 3, "details": "lorem" }
-  ]
-}
