--- ninja-jobs/src/App.vue
@@ -1,65 +1,0 @@
-<template>
-  <div id="nav">
-    <router-link to="/">Home</router-link>
-    <router-link :to="{ name: 'About' }">About</router-link>
-    <router-link :to="{ name: 'Jobs' }">Jobs</router-link>
-  </div>
-
-  <button @click="redirect">Redirect</button>
-  <button @click="back">Go back</button>
-  <button @click="forward">Go forward</button>
-
-  <router-view/>
-</template>
-
-<script>
-export default {
-   methods: {
-    redirect() {
-      this.$router.push({ name: 'Home' })
-    },
-    back() {
-      this.$router.go(-1)
-    },
-    forward() {
-      this.$router.go(1)
-    }
-  }
-}
-</script>
-
-<style>
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-  max-width: 960px;
-  margin: 0 auto;
-}
-
-#nav {
-  padding: 30px;
-}
-
-#nav a {
-  font-weight: bold;
-  color: #2c3e50;
-  text-decoration: none;
-  padding: 10px;
-  border-radius: 4px;
-}
-
-#nav a.router-link-exact-active {
-  color: white;
-  background: crimson;
-}
-
-button {
-  margin: 0 10px;
-  padding: 10px;
-  border: none;
-  border-radius: 4px;
-}
-</style>
