--- ninja-jobs/src/router/index.js
@@ -1,48 +1,0 @@
-import { createRouter, createWebHistory } from 'vue-router'
-import Home from '../views/Home.vue'
-import About from '../views/About.vue'
-import Jobs from '../views//jobs/Jobs.vue'
-import JobDetails from '../views//jobs/JobDetails.vue'
-import NotFound from '../views/NotFound.vue'
-
-const routes = [
-  {
-    path: '/',
-    name: 'Home',
-    component: Home
-  },
-  {
-    path: '/about',
-    name: 'About',
-    component: About
-  },
-  {
-    path: '/jobs',
-    name: 'Jobs',
-    component: Jobs
-  },
-  {
-    path: '/jobs/:id',
-    name: 'JobDetails',
-    component: JobDetails,
-    props: true
-  },
-  // redirect
-  {
-    path: '/all-jobs',
-    redirect: '/jobs'
-  },
-  // 404 catchall
-  {
-    path: '/:catchAll(.*)',
-    name: NotFound,
-    component: () => import('../views/NotFound.vue')
-  }
-]
-
-const router = createRouter({
-  history: createWebHistory(process.env.BASE_URL),
-  routes
-})
-
-export default router
