--- ninja-jobs/src/views/About.vue
@@ -1,7 +1,0 @@
-<template>
-  <div class="about">
-    <h1>This is an about page</h1>
-    <p>Lorem, ipsum dolor sit amet consectetur adipisicing elit. Dignissimos modi sint vel consectetur ut repellendus magnam esse explicabo corrupti ab cum praesentium, iusto culpa nihil optio saepe maiores est non perferendis excepturi at eos. Dolorum dolore molestiae hic similique doloribus saepe pariatur, quos, possimus voluptatem, aliquid aut non necessitatibus quaerat.</p>
-    <p>Lorem, ipsum dolor sit amet consectetur adipisicing elit. Dignissimos modi sint vel consectetur ut repellendus magnam esse explicabo corrupti ab cum praesentium, iusto culpa nihil optio saepe maiores est non perferendis excepturi at eos. Dolorum dolore molestiae hic similique doloribus saepe pariatur, quos, possimus voluptatem, aliquid aut non necessitatibus quaerat.</p>
-  </div>
-</template>
