--- ninja-jobs/src/views/Home.vue
@@ -1,14 +1,0 @@
-<template>
-  <div class="home">
-    <h1>Homepage</h1>
-    <p>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Nam, praesentium dolorum libero reprehenderit fugiat exercitationem necessitatibus! Incidunt, eum. Aperiam facere magnam consectetur omnis molestiae vero. Placeat illum animi deserunt in illo, amet perspiciatis corrupti ex quo aut, ducimus earum sapiente! Odio pariatur facere voluptatibus libero magnam possimus quo voluptatum nobis.</p>
-    <p>Lorem ipsum, dolor sit amet consectetur adipisicing elit. Nam, praesentium dolorum libero reprehenderit fugiat exercitationem necessitatibus! Incidunt, eum. Aperiam facere magnam consectetur omnis molestiae vero. Placeat illum animi deserunt in illo, amet perspiciatis corrupti ex quo aut, ducimus earum sapiente! Odio pariatur facere voluptatibus libero magnam possimus quo voluptatum nobis.</p>
-  </div>
-</template>
-
-<script>
-export default {
-  name: 'Home',
-  components: {}
-}
-</script>
