--- ninja-jobs/src/views/jobs/JobDetails.vue
@@ -1,30 +1,0 @@
-<template>
-  <div v-if="job">
-    <h1>{{ job.title }}</h1>
-    <p>{{ job.details }}</p>
-  </div>
-  <div v-else>
-    <p>Loading job details...</p>
-  </div>
-</template>
-
-<script>
-export default {
-  props: ['id'],
-  data() {
-    return {
-      job: null
-    }
-  },
-  mounted() {
-    fetch('http://localhost:3000/jobs/' + this.id)
-      .then(res => res.json())
-      .then(data => this.job = data)
-      .catch(err => console.log(err.message))
-  }
-}
-</script>
-
-<style>
-
-</style>
