+++ project-planner/README.md
@@ -1,4 +1,4 @@
-# ninja-jobs
+# project-planner
 
 ## Project setup
 ```
