+++ project-planner/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ninja-jobs",
+  "name": "project-planner",
   "version": "0.1.0",
   "private": true,
   "scripts": {
