+++ project-planner/src/assets/logo.png
