+++ project-planner/src/views/Home.vue
@@ -1,0 +1,13 @@
+<template>
+  <div class="home">
+    Home
+  </div>
+</template>
+
+<script>
+
+export default {
+  name: 'Home',
+  components: {}
+}
+</script>
